DataProcessor/api_flask_env/requirements.txt
alembic==1.11.1
APIFlask==1.3.1
apispec==6.3.0
blinker==1.6.2
click==8.1.6
Flask==2.3.2
Flask-Cors==4.0.0
Flask-HTTPAuth==4.8.0
flask-marshmallow==0.15.0
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
mysql-connector-python==8.1.0
packaging==23.1
protobuf==4.21.12
python-dotenv==1.0.0
SQLAlchemy==2.0.19
typing_extensions==4.7.1
webargs==8.3.0
Werkzeug==2.3.6
zipp==3.16.2

mysql-connector-python==8.1.0
